requirements.txt
python-fasthtml
pandas
numpy
plotly